# openapi/pdf-worker.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/pdf/admin/templates": {
      "get": {
        "description": "`origin` distingue los templates incorporados (viajan con el código) de los publicados por la API. Sólo estos últimos se pueden modificar o retirar.",
        "operationId": "PdfTemplateAdminController_inventory",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Credencial de administración ausente o inválida."
          },
          "404": {
            "description": "La administración está desactivada."
          }
        },
        "summary": "Inventario completo, con origen y estado de cada versión",
        "tags": [
          "pdf-templates"
        ]
      },
      "post": {
        "description": "Valida el paquete completo —formato, plantilla, estilos, contrato y datos de ejemplo—, lo compila y lo registra. Si la versión ya existe responde 409 sugiriendo la siguiente: las versiones publicadas son inmutables, porque un informe archivado declara con cuál salió. Actualizar un template es publicar otra versión.",
        "operationId": "PdfTemplateAdminController_publish",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Paquete de template. Vea GET /pdf/template-format/example.",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template publicado y registrado."
          },
          "409": {
            "description": "Esa versión ya está publicada."
          },
          "422": {
            "description": "El paquete no cumple el formato."
          }
        },
        "summary": "Publica un template nuevo",
        "tags": [
          "pdf-templates"
        ]
      }
    },
    "/pdf/admin/templates/{templateId}/{version}/source": {
      "get": {
        "description": "Devuelve el JSON tal cual se subió: es el punto de partida para publicar la versión siguiente. Un template incorporado responde 403 — se edita en el repositorio.",
        "operationId": "PdfTemplateAdminController_source",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "certificado-de-cuenta",
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "example": "1.0.0",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Descarga el paquete de un template publicado por la API",
        "tags": [
          "pdf-templates"
        ]
      }
    },
    "/pdf/admin/templates/{templateId}/{version}/deprecate": {
      "post": {
        "description": "Es la forma recomendada de «borrar»: deja de recomendarse pero SIGUE generando, así que lo ya emitido con ella se puede reproducir.",
        "operationId": "PdfTemplateAdminController_deprecate",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Marca una versión como obsoleta",
        "tags": [
          "pdf-templates"
        ]
      }
    },
    "/pdf/admin/templates/{templateId}/{version}": {
      "delete": {
        "description": "Borrado REAL. Existe para deshacer una publicación equivocada. Si ya se emitieron documentos con esa versión, se pierde la capacidad de reproducirlos: use «deprecate».",
        "operationId": "PdfTemplateAdminController_remove",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Versión retirada del catálogo y del almacén."
          },
          "403": {
            "description": "Es un template incorporado."
          }
        },
        "summary": "Borra una versión publicada por la API",
        "tags": [
          "pdf-templates"
        ]
      }
    }
  },
  "info": {
    "title": "ATLAS PDF Generator Worker",
    "description": "Plataforma documental interna: los artefactos entregan datos estructurados y este servicio valida, maqueta, aplica la identidad institucional y devuelve el PDF.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "admin": {
        "type": "apiKey",
        "in": "header",
        "name": "x-pdf-admin-key"
      }
    },
...
  }
}